requirements.txt
pytesseract==0.3.10
Pillow==10.0.0
python-docx==0.8.11
tk   